+++ src/com/pinktwins/elephant/data/Settings.java
@@ -18,7 +18,7 @@ public class Settings {
 	public static enum Keys {
 		DEFAULT_NOTEBOOK("defaultNotebook"), VAULT_FOLDER("noteFolder"), USE_LUCENE("useLucene"), NOTELIST_MODE("noteListMode"), AUTOBULLET("autoBullet"), RECENT_SHOW(
 				"showRecent"), ALLOW_FILENAMECHARS("allowFilenameChars"), CONFIRM_DELETE_FROM_TRASH("confirmDeleteFromTrash"), WINDOW_MAXIMIZED("maximized"), FONT_SCALE(
-				"fontScale"), PASTE_PLAINTEXT("pastePlaintext"), SHOW_SIDEBAR("showSidebar");
+				"fontScale"), PASTE_PLAINTEXT("pastePlaintext"), SHOW_SIDEBAR("showSidebar"), DEFAULT_FILETYPE("defaultFiletype");
 
 		private final String str;
 
@@ -224,5 +224,12 @@ public class Settings {
 			return true;
 		}
 		return getBoolean(Keys.SHOW_SIDEBAR);
+	}
+	
+	public String getDefaultFiletype() {
+		if (!has(Keys.DEFAULT_FILETYPE)) {
+			return "txt";
+		}
+		return getString(Keys.DEFAULT_FILETYPE);
 	}
 }